Fixed analyzer violations in mykitchen.common project
