+++ src/.editorconfig
@@ -163,7 +163,7 @@ csharp_space_around_declaration_statements = false
 csharp_space_before_colon_in_inheritance_clause = true
 csharp_space_before_comma = false
 csharp_space_before_dot = false
-csharp_space_before_open_square_brackets = false
+csharp_space_before_open_square_brackets = true
 csharp_space_before_semicolon_in_for_statement = false
 csharp_space_between_empty_square_brackets = false
 csharp_space_between_method_call_empty_parameter_list_parentheses = false
