+++ src/Common/MyKitchen.Common/Guard/Guard.cs
@@ -1,9 +1,20 @@
+// |-----------------------------------------------------------------------------------------------------|
+// <copyright file="Guard.cs" company="MyKitchen">
+// Copyright (c) MyKitchen. All Rights Reserved.
+// Licensed under the GPLv3 license. See LICENSE file in the project root for full license information.
+// </copyright>
+// |-----------------------------------------------------------------------------------------------------|
+
 namespace MyKitchen.Common.Guard
 {
     using System.Text.RegularExpressions;
 
+    /// <summary>
+    /// This class implements the <see cref="IGuard"/> interface.
+    /// </summary>
     public class Guard : IGuard
     {
+        /// <inheritdoc/>
         public TOut? AgainstNull<TOut>(object? variable, string formatString, params object?[] args)
         {
             if (variable == null)
@@ -14,6 +25,7 @@ namespace MyKitchen.Common.Guard
             return default;
         }
 
+        /// <inheritdoc/>
         public TOut? AgainstTrue<TOut>(bool boolean, string formatString, params object?[] args)
         {
             if (boolean == true)
@@ -24,6 +36,7 @@ namespace MyKitchen.Common.Guard
             return default;
         }
 
+        /// <inheritdoc/>
         public TOut? AgainstFalse<TOut>(bool boolean, string formatString, params object?[] args)
         {
             if (boolean == false)
@@ -34,6 +47,7 @@ namespace MyKitchen.Common.Guard
             return default;
         }
 
+        /// <inheritdoc/>
         public TOut? AgainstNullOrEmpty<TOut>(string text, string formatString, params object?[] args)
         {
             if (string.IsNullOrEmpty(text))
@@ -44,6 +58,7 @@ namespace MyKitchen.Common.Guard
             return default;
         }
 
+        /// <inheritdoc/>
         public TOut? AgainstNullOrWhiteSpace<TOut>(string text, string formatString, params object?[] args)
         {
             if (string.IsNullOrWhiteSpace(text))
@@ -54,6 +69,7 @@ namespace MyKitchen.Common.Guard
             return default;
         }
 
+        /// <inheritdoc/>
         public TOut? AgainstRegex<TOut>(string text, Regex regex, string formatString, params object?[] args)
         {
             if (regex.IsMatch(text) == false)
@@ -64,6 +80,7 @@ namespace MyKitchen.Common.Guard
             return default;
         }
 
+        /// <inheritdoc/>
         public TOut? AgainstRegex<TOut>(string text, string pattern, string formatString, params object?[] args)
         {
             if (Regex.IsMatch(text, pattern) == false)
